+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/dump/DesensitizedSQLBuilder.java
@@ -52,7 +52,6 @@ import com.starrocks.qe.SessionVariable;
 import com.starrocks.server.CatalogMgr;
 import com.starrocks.sql.analyzer.Analyzer;
 import com.starrocks.sql.analyzer.AnalyzerUtils;
-import com.starrocks.sql.analyzer.AstToSQLBuilder;
 import com.starrocks.sql.analyzer.Field;
 import com.starrocks.sql.analyzer.QueryAnalyzer;
 import com.starrocks.sql.ast.CTERelation;
@@ -69,6 +68,7 @@ import com.starrocks.sql.ast.TableRelation;
 import com.starrocks.sql.ast.ValuesRelation;
 import com.starrocks.sql.ast.ViewRelation;
 import com.starrocks.sql.common.MetaUtils;
+import com.starrocks.sql.formatter.AST2SQLVisitor;
 import com.starrocks.sql.parser.SqlParser;
 import com.starrocks.statistic.StatsConstants;
 import org.apache.commons.collections4.CollectionUtils;
@@ -100,7 +100,7 @@ public class DesensitizedSQLBuilder {
     public static String desensitizeSQL(StatementBase statement, Map<String, String> desensitizedDict) {
         Map<TableName, Table> tables = AnalyzerUtils.collectAllTableAndViewWithAlias(statement);
         boolean sameCatalogDb = tables.keySet().stream().map(TableName::getCatalogAndDb).distinct().count() == 1;
-        return new DesensitizedSQLVisitor(sameCatalogDb, false, desensitizedDict).visit(statement);
+        return new DesensitizedSQLVisitor(sameCatalogDb, true, desensitizedDict).visit(statement);
     }
 
     public static String desensitizeViewDef(View view, Map<String, String> desensitizedDict, ConnectContext connectContext) {
@@ -108,15 +108,15 @@ public class DesensitizedSQLBuilder {
         new QueryAnalyzer(connectContext).analyze(stmt);
         Map<TableName, Table> tables = AnalyzerUtils.collectAllTableAndViewWithAlias(stmt);
         boolean sameCatalogDb = tables.keySet().stream().map(TableName::getCatalogAndDb).distinct().count() == 1;
-        return new DesensitizedSQLVisitor(sameCatalogDb, false, desensitizedDict).desensitizeViewDef(stmt);
+        return new DesensitizedSQLVisitor(sameCatalogDb, true, desensitizedDict).desensitizeViewDef(stmt);
     }
 
     public static String desensitizeTableDef(Pair<String, Table> pair, Map<String, String> desensitizedDict) {
         Table table = pair.second;
-        DesensitizedSQLVisitor visitor = new DesensitizedSQLVisitor(true, true, desensitizedDict);
+        DesensitizedSQLVisitor visitor = new DesensitizedSQLVisitor(true, false, desensitizedDict);
         String tableDef = "";
         if (table.isMaterializedView()) {
-            visitor = new DesensitizedSQLVisitor(true, false, desensitizedDict);
+            visitor = new DesensitizedSQLVisitor(true, true, desensitizedDict);
             tableDef = visitor.desensitizeMvDef(table);
         } else if (table.getType() == Table.TableType.MYSQL || table.getType() == Table.TableType.ELASTICSEARCH
                 || table.getType() == Table.TableType.BROKER || table.getType() == Table.TableType.HIVE
@@ -151,12 +151,13 @@ public class DesensitizedSQLBuilder {
         return desensitizedDict.get(colName);
     }
 
-    public static class DesensitizedSQLVisitor extends AstToSQLBuilder.AST2SQLBuilderVisitor {
+    public static class DesensitizedSQLVisitor extends AST2SQLVisitor {
 
         private final Map<String, String> desensitizedDict;
 
-        public DesensitizedSQLVisitor(boolean simple, boolean withoutTbl, Map<String, String> desensitizedDict) {
-            super(simple, withoutTbl, true);
+        public DesensitizedSQLVisitor(boolean simple, boolean withTbl, Map<String, String> desensitizedDict) {
+            options.setColumnSimplifyTableName(simple);
+            options.setColumnWithTableName(withTbl);
             this.desensitizedDict = desensitizedDict;
         }
 
@@ -608,7 +609,7 @@ public class DesensitizedSQLBuilder {
 
             sb.append("\n)");
             String define = materializedView.getSimpleDefineSql();
-            if (StringUtils.isEmpty(define) || !simple) {
+            if (StringUtils.isEmpty(define) || !options.isColumnSimplifyTableName()) {
                 define = materializedView.getViewDefineSql();
             }
             StatementBase stmt = SqlParser.parse(define, SessionVariable.DEFAULT_SESSION_VARIABLE).get(0);
@@ -826,8 +827,8 @@ public class DesensitizedSQLBuilder {
 
         private String desensitizeColumnName(TableName tableName, String fieldName, String aliasName) {
             String res = "";
-            if (tableName != null && !withoutTbl) {
-                if (!simple) {
+            if (tableName != null && options.isColumnWithTableName()) {
+                if (!options.isColumnSimplifyTableName()) {
                     res = desensitizeTableName(tableName);
                 } else {
                     res = "tbl_" + desensitizeValue(tableName.getTbl(), "table");
