+++ fe/fe-core/src/test/java/com/starrocks/sql/common/SqlDigestBuilderTest.java
@@ -14,6 +14,8 @@
 
 package com.starrocks.sql.common;
 
+import com.google.common.collect.Lists;
+import com.starrocks.sql.analyzer.AstToSQLBuilder;
 import com.starrocks.sql.ast.StatementBase;
 import com.starrocks.sql.plan.PlanTestBase;
 import com.starrocks.utframe.UtFrameUtils;
@@ -22,7 +24,10 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+
+import java.util.List;
 
 public class SqlDigestBuilderTest extends PlanTestBase {
 
@@ -36,56 +41,65 @@ public class SqlDigestBuilderTest extends PlanTestBase {
         PlanTestBase.afterClass();
     }
 
+    private static List<Arguments> source() {
+        List<Arguments> list = Lists.newArrayList();
+        list.add(Arguments.of("select * from t2", "SELECT * FROM `test`.`t2`"));
+        list.add(Arguments.of("select * from t1 where v4 = 1 and v5 = 2",
+                "SELECT * FROM `test`.`t1` WHERE (`v4` = ? AND `v5` = ?)"));
+        list.add(Arguments.of("select * from t1 where v4 = 1 or v5 = 2",
+                "SELECT * FROM `test`.`t1` WHERE (`v4` = ? OR `v5` = ?)"));
+        list.add(Arguments.of("select * from t1 where v4 = 1 and (v5 = 2 or v6 = 3)",
+                "SELECT * FROM `test`.`t1` WHERE (`v4` = ? AND (`v5` = ? OR `v6` = ?))"));
+        list.add(Arguments.of("select v4 from t1 limit 1", "SELECT `v4` FROM `test`.`t1` LIMIT  ?"));
+        list.add(Arguments.of("select * from t1 where v4 in (1, 2, 3)",
+                "SELECT * FROM `test`.`t1` WHERE `v4` IN (?, ?, ?)"));
+        list.add(Arguments.of("select * from t1 where v4 in (select v5 from t1)",
+                "SELECT * FROM `test`.`t1` WHERE `v4` IN (((SELECT `v5` FROM `test`.`t1`)))"));
+        list.add(Arguments.of("select /*+set_var(query_timeout=123)*/ * from t2",
+                "SELECT /*+set_var(query_timeout=123)*/ * FROM `test`.`t2`"));
+        list.add(Arguments.of("insert into t1 values (1, 2, 3)", "INSERT INTO `test`.`t1` VALUES(?, ?, ?)"));
+        list.add(Arguments.of("insert into t1 values (1, 2, 3),(4,5,6)", "INSERT INTO `test`.`t1` VALUES(?, ?, ?)"));
+        list.add(Arguments.of("insert into t1 select * from t1", "INSERT INTO `test`.`t1` SELECT * FROM `test`.`t1`"));
+        list.add(Arguments.of("insert into t1 with label abc select * from t1",
+                "INSERT INTO `test`.`t1` WITH LABEL ? SELECT * FROM `test`.`t1`"));
+        list.add(Arguments.of("delete from t1 where v4 = 1", "DELETE FROM `test`.`t1` WHERE `v4` = ?"));
+        list.add(Arguments.of("insert into part_t1 partition (p1) values(1,2,3)",
+                "INSERT INTO `test`.`part_t1` PARTITION (p1) VALUES(?, ?, ?)"));
+        list.add(Arguments.of("insert overwrite part_t1 partition (p1) values(1,2,3)",
+                "INSERT OVERWRITE `test`.`part_t1` PARTITION (p1) VALUES(?, ?, ?)"));
+        list.add(Arguments.of(
+                "select * from t1 where v4=1 or v4=2 or v4=3 or v4=4 or v4=5 or v4=6 or v4=7 or v4=8 or v4=9 or v4=10"
+                        + " or v4=11 or v4=12 or v4=13 or v4=14 or v4=15 or v4=16 or v4=17 or v4=18 or v4=19 or v4=20",
+                "SELECT * FROM `test`.`t1` WHERE (`v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = "
+                        + "? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR "
+                        + "`v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ?)"));
+        list.add(Arguments.of(
+                "select * from t1 where v4+v5=1 or v4+v5=2 or v4+v5=3 or v4=4 or v4=5 or v4=6 or v4=7 or v4=8 or v4=9"
+                        + " or v4=10 or v4=11 or v4=12 or v4=13 or v4=14 or v4=15 or v4=16 or v4=17 or v4=18 or v4=19"
+                        + " or v4=20",
+                "SELECT * FROM `test`.`t1` WHERE ((`v4` + `v5`) = ? OR (`v4` + `v5`) = ? OR (`v4` + `v5`) = ? OR `v4`"
+                        + " = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? "
+                        + "OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR "
+                        + "`v4` = ? OR `v4` = ?)"));
+        list.add(Arguments.of(
+                "select * from t1 where v5 = 123 and (v4=1 or v4=2 or v4=3 or v4=4 or v4=5 or v4=6 or v4=7 or v4=8 or"
+                        + " v4=9 or v4=10 or v4=11 or v4=12 or v4=13 or v4=14 or v4=15 or v4=16 or v4=17 or v4=18 or "
+                        + "v4=19 or v4=20)",
+                "SELECT * FROM `test`.`t1` WHERE (`v5` = ? AND (`v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` "
+                        + "= ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR"
+                        + " `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` = ? OR `v4` "
+                        + "= ?))"));
+        return list;
+    }
+
     @ParameterizedTest
-    @CsvSource(delimiterString = "|", value = {
-            "select * from t2| SELECT * FROM test.t2",
-            "select * from t1 where v4 = 1 and v5 = 2|SELECT * FROM test.t1 WHERE (test.t1.v4 = ?) AND (test.t1.v5 = " +
-                    "?)",
-            "select * from t1 where v4 = 1 or v5 = 2|SELECT * FROM test.t1 WHERE (test.t1.v4 = ?) OR (test.t1.v5 = ?)",
-            "select * from t1 where v4 = 1 and (v5 = 2 or v6 = 3)| SELECT * FROM test.t1 WHERE (test.t1.v4 = ?) AND (" +
-                    "(test.t1.v5 = ?) OR (test.t1.v6 = ?))",
-            "select v4 from t1 limit 1|SELECT test.t1.v4 FROM test.t1 LIMIT  ? ",
-            "select * from t1 where v4 in (1, 2, 3) | SELECT * FROM test.t1 WHERE test.t1.v4 IN (?)",
-            "select * from t1 where v4 in (select v5 from t1)| SELECT * FROM test.t1 WHERE test.t1.v4 IN (((SELECT " +
-                    "test.t1.v5 FROM test.t1)))",
-            // with set_var
-            "select /*+set_var(query_timeout=123)*/ * from t2| SELECT * FROM test.t2",
-
-            // insert
-            "insert into t1 values (1, 2, 3)| INSERT INTO `test`.`t1` VALUES(?, ?, ?)",
-            "insert into t1 values (1, 2, 3),(4,5,6)| INSERT INTO `test`.`t1` VALUES(?, ?, ?)",
-            "insert into t1 select * from t1| INSERT INTO `test`.`t1` SELECT * FROM test.t1",
-            "insert into t1 with label abc select * from t1| " +
-                    "INSERT INTO `test`.`t1` WITH LABEL ? SELECT * FROM test.t1",
-
-            // delete
-            "delete from t1 where v4 = 1|DELETE FROM `test`.`t1` WHERE v4 = ?",
-
-            // partition dml
-            "insert into part_t1 partition (p1) values(1,2,3) " +
-                    "|INSERT INTO `test`.`part_t1` PARTITION (p1) VALUES(?, ?, ?)",
-            "insert overwrite part_t1 partition (p1) values(1,2,3) " +
-                    "|INSERT OVERWRITE `test`.`part_t1` PARTITION (p1) VALUES(?, ?, ?)",
-
-            // massive compounds
-            "select * from t1 where v4=1 or v4=2 or v4=3 or v4=4 or v4=5 or v4=6 or v4=7 or v4=8 or v4=9 or v4=10 " +
-                    "or v4=11 or v4=12 or v4=13 or v4=14 or v4=15 or v4=16 or v4=17 or v4=18 " +
-                    "or v4=19 or v4=20| " +
-                    "SELECT * FROM test.t1 WHERE $massive_compounds[`test`.`t1`.`v4`]$",
-            "select * from t1 where v4+v5=1 or v4+v5=2 or v4+v5=3 or v4=4 or v4=5 or v4=6 or v4=7 or v4=8 or v4=9 or " +
-                    "v4=10 " +
-                    "or v4=11 or v4=12 or v4=13 or v4=14 or v4=15 or v4=16 or v4=17 or v4=18 " +
-                    "or v4=19 or v4=20| " +
-                    "SELECT * FROM test.t1 WHERE $massive_compounds[`test`.`t1`.`v4`,`test`.`t1`.`v5`]$",
-            "select * from t1 where v5 = 123 and (v4=1 or v4=2 or v4=3 or v4=4 or v4=5 or v4=6 or v4=7 or v4=8 or " +
-                    "v4=9 or v4=10 " +
-                    "or v4=11 or v4=12 or v4=13 or v4=14 or v4=15 or v4=16 or v4=17 or v4=18 " +
-                    "or v4=19 or v4=20)| " +
-                    "SELECT * FROM test.t1 WHERE $massive_compounds[`test`.`t1`.`v4`,`test`.`t1`.`v5`]$",
-    })
+    @MethodSource("source")
     public void testBuild(String sql, String expectedDigest) throws Exception {
         StatementBase stmt = UtFrameUtils.parseStmtWithNewParser(sql, connectContext);
-        Assertions.assertEquals(StringUtils.trim(expectedDigest), StringUtils.trim(SqlDigestBuilder.build(stmt)));
+        String cc = "list.add(Arguments.of(\"" + sql + "\", \"" + StringUtils.trim(AstToSQLBuilder.toDigest(stmt))
+                + "\"));";
+        cc = cc.replace("\n", "\\n");
+        Assertions.assertEquals(StringUtils.trim(expectedDigest), StringUtils.trim(AstToSQLBuilder.toDigest(stmt)));
     }
 
 }
