+++ fe/fe-core/src/test/java/com/starrocks/sql/spm/SPMPlanBindTest.java
@@ -167,14 +167,12 @@ public class SPMPlanBindTest extends PlanTestBase {
                         " t1 ON v3 = v6");
         SPMPlanBuilder generator = new SPMPlanBuilder(connectContext, stmt);
         generator.execute();
-        assertContains(generator.getBindSqlDigest(), "SELECT `test`.`t1`.`v4`, `test`.`t0`.`v2` " +
-                "FROM `test`.`t0` , `test`.`t1`  " +
-                "WHERE (`test`.`t0`.`v3` = `test`.`t1`.`v6`) AND (`test`.`t0`.`v2` IN (?))");
+        assertContains(generator.getBindSqlDigest(), "SELECT `test`.`t1`.`v4`, `test`.`t0`.`v2` FROM `test`.`t0` , `test`.`t1`  "
+                + "WHERE (`test`.`t0`.`v3` = `test`.`t1`.`v6` AND `test`.`t0`.`v2` IN (?))");
 
-        assertContains(generator.getBindSql(), "SELECT `test`.`t1`.`v4`, `test`.`t0`.`v2` " +
-                "FROM `test`.`t0` , `test`.`t1`  " +
-                "WHERE (`test`.`t0`.`v3` = `test`.`t1`.`v6`) " +
-                "AND (`test`.`t0`.`v2` IN (_spm_const_list(1, 1, 2, 3, 4)))");
+        assertContains(generator.getBindSql(), "SELECT `test`.`t1`.`v4`, `test`.`t0`.`v2` "
+                + "FROM `test`.`t0` , `test`.`t1`  WHERE (`test`.`t0`.`v3` = `test`.`t1`.`v6` "
+                + "AND `test`.`t0`.`v2` IN (_spm_const_list(1, 1, 2, 3, 4)))");
 
         assertContains(generator.getPlanStmtSQL(),
                 "SELECT v2, v4 FROM (SELECT * FROM t0 WHERE v2 IN (_spm_const_list(1, 1, 2, 3, 4))) t_0 "
@@ -193,15 +191,13 @@ public class SPMPlanBindTest extends PlanTestBase {
         SPMPlanBuilder generator = new SPMPlanBuilder(connectContext, stmt);
         generator.execute();
         assertContains(generator.getBindSqlDigest(), "SELECT `test`.`t1`.`v4`, `test`.`t0`.`v2` "
-                + "FROM `test`.`t0` , `test`.`t1`  "
-                + "WHERE ((`test`.`t0`.`v3` = `test`.`t1`.`v6`) "
-                + "AND (`test`.`t0`.`v2` IN (?))) AND (`test`.`t1`.`v5` = ?)");
+                + "FROM `test`.`t0` , `test`.`t1`  WHERE (`test`.`t0`.`v3` = `test`.`t1`.`v6` "
+                + "AND `test`.`t0`.`v2` IN (?) AND `test`.`t1`.`v5` = ?)");
 
         assertContains(generator.getBindSql(), "SELECT `test`.`t1`.`v4`, `test`.`t0`.`v2` "
-                + "FROM `test`.`t0` , `test`.`t1`  "
-                + "WHERE ((`test`.`t0`.`v3` = `test`.`t1`.`v6`) "
-                + "AND (`test`.`t0`.`v2` IN (_spm_const_list(1, 10, 11)))) "
-                + "AND (`test`.`t1`.`v5` = _spm_const_var(2, 5))");
+                + "FROM `test`.`t0` , `test`.`t1`  WHERE (`test`.`t0`.`v3` = `test`.`t1`.`v6` "
+                + "AND `test`.`t0`.`v2` IN (_spm_const_list(1, 10, 11)) "
+                + "AND `test`.`t1`.`v5` = _spm_const_var(2, 5))");
 
         assertContains(generator.getPlanStmtSQL(),
                 "SELECT v2, v4 FROM (SELECT * FROM t0 WHERE v2 IN (_spm_const_list(1, 10, 11))) t_0 "
@@ -221,15 +217,13 @@ public class SPMPlanBindTest extends PlanTestBase {
         SPMPlanBuilder generator = new SPMPlanBuilder(connectContext, stmt);
         generator.execute();
         assertContains(generator.getBindSqlDigest(), "SELECT `test`.`t1`.`v4`, `test`.`t0`.`v2` "
-                + "FROM `test`.`t0` , `test`.`t1`  "
-                + "WHERE ((`test`.`t0`.`v3` = `test`.`t1`.`v6`) "
-                + "AND (`test`.`t0`.`v2` IN (?))) AND (`test`.`t1`.`v5` = ?)");
+                + "FROM `test`.`t0` , `test`.`t1`  WHERE (`test`.`t0`.`v3` = `test`.`t1`.`v6` "
+                + "AND `test`.`t0`.`v2` IN (?) AND `test`.`t1`.`v5` = ?)");
 
         assertContains(generator.getBindSql(), "SELECT `test`.`t1`.`v4`, `test`.`t0`.`v2` "
-                + "FROM `test`.`t0` , `test`.`t1`  "
-                + "WHERE ((`test`.`t0`.`v3` = `test`.`t1`.`v6`) "
-                + "AND (`test`.`t0`.`v2` IN (_spm_const_list(1, 10, 11)))) "
-                + "AND (`test`.`t1`.`v5` = _spm_const_var(2, 1))");
+                + "FROM `test`.`t0` , `test`.`t1`  WHERE (`test`.`t0`.`v3` = `test`.`t1`.`v6` "
+                + "AND `test`.`t0`.`v2` IN (_spm_const_list(1, 10, 11)) "
+                + "AND `test`.`t1`.`v5` = _spm_const_var(2, 1))");
 
         assertContains(generator.getPlanStmtSQL(), "SELECT v2, v4 FROM "
                 + "(SELECT * FROM t0 WHERE v2 IN (_spm_const_list(1, 10, 11))) t_0 "
